+++ src/platform/linux/x11/mod.rs
@@ -14,6 +14,8 @@ use events::ModifiersState;
 use std::{mem, ptr, slice};
 use std::sync::{Arc, Mutex, Weak};
 use std::sync::atomic::{self, AtomicBool};
+use std::sync::mpsc::{channel, Receiver, Sender};
+use std::cell::RefCell;
 use std::collections::HashMap;
 use std::ffi::CStr;
 use std::os::raw::{c_char, c_int, c_long, c_uchar, c_ulong};
@@ -40,6 +42,9 @@ pub struct EventsLoop {
     display: Arc<XConnection>,
     wm_delete_window: ffi::Atom,
     dnd: Dnd,
+    ime_receiver: Receiver<(WindowId, i16, i16)>,
+    ime_sender: Sender<(WindowId, i16, i16)>,
+    ime: RefCell<HashMap<WindowId, util::Ime>>,
     windows: Arc<Mutex<HashMap<WindowId, WindowData>>>,
     devices: Mutex<HashMap<DeviceId, Device>>,
     xi2ext: XExtension,
@@ -64,6 +69,8 @@ impl EventsLoop {
 
         let dnd = Dnd::new(Arc::clone(&display))
             .expect("Failed to call XInternAtoms when initializing drag and drop");
+
+        let (ime_sender, ime_receiver) = channel();
 
         let xi2ext = unsafe {
             let mut result = XExtension {
@@ -106,6 +113,9 @@ impl EventsLoop {
             display,
             wm_delete_window,
             dnd,
+            ime_receiver,
+            ime_sender,
+            ime: RefCell::new(HashMap::new()),
             windows: Arc::new(Mutex::new(HashMap::new())),
             devices: Mutex::new(HashMap::new()),
             xi2ext,
@@ -219,17 +229,10 @@ impl EventsLoop {
                 if client_msg.data.get_long(0) as ffi::Atom == self.wm_delete_window {
                     callback(Event::WindowEvent { window_id, event: WindowEvent::Closed });
 
-                    let mut windows = self.windows.lock().unwrap();
-                    let window_data = windows.remove(&WindowId(window));
-                    let _lock = GLOBAL_XOPENIM_LOCK.lock().unwrap();
-                    unsafe {
-                        if let Some(window_data) = window_data {
-                            (self.display.xlib.XDestroyIC)(window_data.ic);
-                            (self.display.xlib.XCloseIM)(window_data.im);
-                            self.display.check_errors()
-                                .expect("Failed to close XIM");
+                    if let Some(_) = self.windows.lock().unwrap().remove(&WindowId(window)) {
+                        unsafe {
+                            (self.display.xlib.XDestroyWindow)(self.display.display, window);
                         }
-                        (self.display.xlib.XDestroyWindow)(self.display.display, window);
                         self.display.check_errors()
                             .expect("Failed to destroy window");
                     }
@@ -405,6 +408,14 @@ impl EventsLoop {
                 }
             }
 
+            ffi::DestroyNotify => {
+                let xev: &ffi::XDestroyWindowEvent = xev.as_ref();
+
+                let window = xev.window;
+
+                self.ime.borrow_mut().remove(&WindowId(window));
+            }
+
             ffi::Expose => {
                 let xev: &ffi::XExposeEvent = xev.as_ref();
 
@@ -458,36 +469,33 @@ impl EventsLoop {
 
                 if state == Pressed {
                     let written = unsafe {
-                        use std::str;
+                        if let Some(ime) = self.ime.borrow().get(&WindowId(window)) {
+                            use std::str;
 
-                        const INIT_BUFF_SIZE: usize = 16;
-                        let mut windows = self.windows.lock().unwrap();
-                        let window_data = {
-                            if let Some(window_data) = windows.get_mut(&WindowId(window)) {
-                                window_data
-                            } else {
-                                return;
+                            const INIT_BUFF_SIZE: usize = 16;
+
+                            /* buffer allocated on heap instead of stack, due to the possible
+                             * reallocation */
+                            let mut buffer: Vec<u8> = vec![mem::uninitialized(); INIT_BUFF_SIZE];
+                            let mut keysym: ffi::KeySym = 0;
+                            let mut status: ffi::Status = 0;
+                            let mut count = (self.display.xlib.Xutf8LookupString)(ime.ic, xkev,
+                                                                              mem::transmute(buffer.as_mut_ptr()),
+                                                                              buffer.len() as libc::c_int,
+                                                                              &mut keysym, &mut status);
+                            /* buffer overflowed, dynamically reallocate */
+                            if status == ffi::XBufferOverflow {
+                                buffer = vec![mem::uninitialized(); count as usize];
+                                count = (self.display.xlib.Xutf8LookupString)(ime.ic, xkev,
+                                                                              mem::transmute(buffer.as_mut_ptr()),
+                                                                              buffer.len() as libc::c_int,
+                                                                              &mut keysym, &mut status);
                             }
-                        };
-                        /* buffer allocated on heap instead of stack, due to the possible
-                         * reallocation */
-                        let mut buffer: Vec<u8> = vec![mem::uninitialized(); INIT_BUFF_SIZE];
-                        let mut keysym: ffi::KeySym = 0;
-                        let mut status: ffi::Status = 0;
-                        let mut count = (self.display.xlib.Xutf8LookupString)(window_data.ic, xkev,
-                                                                          mem::transmute(buffer.as_mut_ptr()),
-                                                                          buffer.len() as libc::c_int,
-                                                                          &mut keysym, &mut status);
-                        /* buffer overflowed, dynamically reallocate */
-                        if status == ffi::XBufferOverflow {
-                            buffer = vec![mem::uninitialized(); count as usize];
-                            count = (self.display.xlib.Xutf8LookupString)(window_data.ic, xkev,
-                                                                          mem::transmute(buffer.as_mut_ptr()),
-                                                                          buffer.len() as libc::c_int,
-                                                                          &mut keysym, &mut status);
-                        }
 
-                        str::from_utf8(&buffer[..count as usize]).unwrap_or("").to_string()
+                            str::from_utf8(&buffer[..count as usize]).unwrap_or("").to_string()
+                        } else {
+                            return;
+                        }
                     };
 
                     for chr in written.chars() {
@@ -735,11 +743,8 @@ impl EventsLoop {
 
                         let window_id = mkwid(xev.event);
 
-                        let mut windows = self.windows.lock().unwrap();
-                        if let Some(window_data) = windows.get_mut(&WindowId(xev.event)) {
-                            unsafe {
-                                (self.display.xlib.XSetICFocus)(window_data.ic);
-                            }
+                        if let Some(ime) = self.ime.borrow().get(&WindowId(xev.event)) {
+                            ime.focus().expect("Failed to focus input context");
                         } else {
                             return;
                         }
@@ -766,11 +771,8 @@ impl EventsLoop {
                     ffi::XI_FocusOut => {
                         let xev: &ffi::XIFocusOutEvent = unsafe { &*(xev.data as *const _) };
 
-                        let mut windows = self.windows.lock().unwrap();
-                        if let Some(window_data) = windows.get_mut(&WindowId(xev.event)) {
-                            unsafe {
-                                (self.display.xlib.XUnsetICFocus)(window_data.ic);
-                            }
+                        if let Some(ime) = self.ime.borrow().get(&WindowId(xev.event)) {
+                            ime.unfocus().expect("Failed to unfocus input context");
                         } else {
                             return;
                         }
@@ -890,6 +892,15 @@ impl EventsLoop {
 
             _ => {}
         }
+
+        match self.ime_receiver.try_recv() {
+            Ok((window_id, x, y)) => {
+                if let Some(ime) = self.ime.borrow_mut().get_mut(&window_id) {
+                    ime.send_xim_spot(x, y);
+                }
+            }
+            Err(_) => ()
+        }
     }
 
     fn init_device(&self, device: c_int) {
@@ -983,6 +994,7 @@ pub struct Window {
     pub window: Arc<Window2>,
     display: Weak<XConnection>,
     windows: Weak<Mutex<HashMap<WindowId, WindowData>>>,
+    ime_sender: Sender<(WindowId, i16, i16)>,
 }
 
 impl ::std::ops::Deref for Window {
@@ -993,48 +1005,19 @@ impl ::std::ops::Deref for Window {
     }
 }
 
-// XOpenIM doesn't seem to be thread-safe
-lazy_static! {      // TODO: use a static mutex when that's possible, and put me back in my function
-    static ref GLOBAL_XOPENIM_LOCK: Mutex<()> = Mutex::new(());
-}
-
 impl Window {
-    pub fn new(x_events_loop: &EventsLoop,
-               window: &::WindowAttributes,
-               pl_attribs: &PlatformSpecificWindowBuilderAttributes)
-               -> Result<Self, CreationError>
-    {
-        let win = ::std::sync::Arc::new(try!(Window2::new(&x_events_loop, window, pl_attribs)));
+    pub fn new(
+        x_events_loop: &EventsLoop,
+        window: &::WindowAttributes,
+        pl_attribs: &PlatformSpecificWindowBuilderAttributes
+    ) -> Result<Self, CreationError> {
+        let win = Arc::new(try!(Window2::new(&x_events_loop, window, pl_attribs)));
 
-        // creating IM
-        let im = unsafe {
-            let _lock = GLOBAL_XOPENIM_LOCK.lock().unwrap();
-
-            let im = (x_events_loop.display.xlib.XOpenIM)(x_events_loop.display.display, ptr::null_mut(), ptr::null_mut(), ptr::null_mut());
-            if im.is_null() {
-                panic!("XOpenIM failed");
-            }
-            im
-        };
-
-        // creating input context
-        let ic = unsafe {
-            let ic = (x_events_loop.display.xlib.XCreateIC)(im,
-                                              b"inputStyle\0".as_ptr() as *const _,
-                                              ffi::XIMPreeditNothing | ffi::XIMStatusNothing, b"clientWindow\0".as_ptr() as *const _,
-                                              win.id().0, ptr::null::<()>());
-            if ic.is_null() {
-                panic!("XCreateIC failed");
-            }
-            (x_events_loop.display.xlib.XSetICFocus)(ic);
-            x_events_loop.display.check_errors().expect("Failed to call XSetICFocus");
-            ic
-        };
+        let ime = util::Ime::new(Arc::clone(&x_events_loop.display), win.id().0)
+            .expect("Failed to initialize IME");
+        x_events_loop.ime.borrow_mut().insert(win.id(), ime);
 
         x_events_loop.windows.lock().unwrap().insert(win.id(), WindowData {
-            im: im,
-            ic: ic,
-            ic_spot: ffi::XPoint {x: 0, y: 0},
             config: None,
             multitouch: window.multitouch,
             cursor_pos: None,
@@ -1044,6 +1027,7 @@ impl Window {
             window: win,
             windows: Arc::downgrade(&x_events_loop.windows),
             display: Arc::downgrade(&x_events_loop.display),
+            ime_sender: x_events_loop.ime_sender.clone(),
         })
     }
 
@@ -1054,22 +1038,7 @@ impl Window {
 
     #[inline]
     pub fn send_xim_spot(&self, x: i16, y: i16) {
-        if let (Some(windows), Some(display)) = (self.windows.upgrade(), self.display.upgrade()) {
-            let nspot = ffi::XPoint{x: x, y: y};
-            let mut windows = windows.lock().unwrap();
-            let w = windows.get_mut(&self.window.id()).unwrap();
-            if w.ic_spot.x == x && w.ic_spot.y == y {
-                return
-            }
-            w.ic_spot = nspot;
-            unsafe {
-                let preedit_attr = (display.xlib.XVaCreateNestedList)
-                    (0, b"spotLocation\0", &nspot, ptr::null::<()>());
-                (display.xlib.XSetICValues)(w.ic, b"preeditAttributes\0",
-                                            preedit_attr, ptr::null::<()>());
-                (display.xlib.XFree)(preedit_attr);
-            }
-        }
+        let _ = self.ime_sender.send((self.window.id(), x, y));
     }
 }
 
@@ -1104,9 +1073,6 @@ impl Drop for Window {
 /// State maintained for translating window-related events
 struct WindowData {
     config: Option<WindowConfig>,
-    im: ffi::XIM,
-    ic: ffi::XIC,
-    ic_spot: ffi::XPoint,
     multitouch: bool,
     cursor_pos: Option<(f64, f64)>,
 }